update according to review: move libzkp
