--- common/libzkp/go.mod
@@ -1,3 +1,0 @@
-module scroll-tech/common/libzkp
-
-go 1.22
